Updating Scary Wording about submission
In the previous cohort, many students were afraid to submit anything since it said only submit when you have a high quality solution. This caused many students to not try to submit until right before presentation which caused confusion and lead to some groups not having a score at all.